+++ helpers.go
@@ -49,7 +49,7 @@ func (c *columnDefinition) convert() common.TableColumnDefiner {
 	case *datetimeType:
 		return common.NewDatetimeTableColumn(c.identifier, c.defaultValue(), c.nullable(), c.autoincrementable())
 	case *charType:
-		return common.NewCharTableColumn(c.identifier, c.defaultValue(), c.nullable(), c.autoincrementable(), uint32(v.length))
+		return common.NewCharTableColumn(c.identifier, c.defaultValue(), c.nullable(), c.autoincrementable(), uint16(v.length))
 	}
 
 	return nil
